refactor(create manifest): Move preparing manifest to ArFSDao PE-477
